Bump version to 0.8.0, add respective migration
As this package is split out of securedrop-workstation-dom0-config, and
has undergone a little cleanup in the meantime to be a better desktop
application citizen and to avoid confusion with regards to naming and
purpose, a couple of files are now expected to be in a different
location than they were before.

To make sure that everything is compatible with older installations, we
add a migration that moves everything into the new expected places, and
replaces crontabs with systemd user timers.

(Also a small fix for update_version.sh)